+++ src/main/java/com/example/WarehouseDatabaseJava/model/users/manager/stage/DepartmentService.java
@@ -4,6 +4,7 @@ import com.example.WarehouseDatabaseJava.model.order.Custom;
 import com.example.WarehouseDatabaseJava.model.order.CustomRepository;
 import com.example.WarehouseDatabaseJava.model.users.manager.Manager;
 import com.example.WarehouseDatabaseJava.model.users.manager.ManagerProfileDTO;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -20,16 +21,19 @@ public class DepartmentService {
 
     // збереження нового відділу
     public Department save(String departmentName) {
-        Department existingDepartment = departmentRepository.findByDepartmentName(departmentName);
-        if (existingDepartment != null) {
-            throw new IllegalArgumentException("Department with the same name already exists");
+        if (departmentRepository.existsByDepartmentName(departmentName)) {
+            throw new RuntimeException("Department with name: " + departmentName + " already exists");
         }
         Department department = new Department(departmentName);
         return departmentRepository.save(department);
     }
 
     //видалити відділ по його id
-    public void removeDepartmentById(String departmentId){
+    public void removeDepartmentById(String departmentId) {
+        if (!departmentRepository.existsById(departmentId)) {
+            throw new EntityNotFoundException("Department not found with id: " + departmentId);
+        }
+
         departmentRepository.deleteById(departmentId);
     }
 
@@ -50,19 +54,22 @@ public class DepartmentService {
 
     //метод отримання всіх менеджерів, призначених на певний етап NOT USING IN CLIENT NOW!
     public List<ManagerProfileDTO> getAllManagersForDepartment(String departmentId) {
+        if (!departmentRepository.existsById(departmentId)) {
+            throw new EntityNotFoundException("Department not found with id: " + departmentId);
+        }
+
         Department department = departmentRepository.getReferenceById(departmentId);
         List<ManagerProfileDTO> managerProfiles = new ArrayList<>();
-        if (department != null) {
-            List<Manager> managers = department.getManagerList();
 
-            for (Manager manager : managers) {
-                ManagerProfileDTO managerProfileDTO = new ManagerProfileDTO(
-                        manager.getName(),
-                        manager.getSurname(),
-                        manager.getEmail()
-                );
-                managerProfiles.add(managerProfileDTO);
-            }
+        List<Manager> managers = department.getManagerList();
+
+        for (Manager manager : managers) {
+            ManagerProfileDTO managerProfileDTO = new ManagerProfileDTO(
+                    manager.getName(),
+                    manager.getSurname(),
+                    manager.getEmail()
+            );
+            managerProfiles.add(managerProfileDTO);
         }
         return managerProfiles;
     }
@@ -70,15 +77,15 @@ public class DepartmentService {
     //метод призначення конкретному замовленню конкретного відділу доставки
     //ОНОВЛЕНО!!!
     public void assignDepartmentToCustom(String customId, String departmentId) {
-        Custom custom = customRepository.getReferenceById(customId);
-        if (custom == null) {
-            throw new RuntimeException("Custom not found with id: " + customId);
+        if (!customRepository.existsById(customId)) {
+            throw new IllegalArgumentException("Custom not found with id: " + customId);
+        }
+        if (!departmentRepository.existsById(departmentId)) {
+            throw new EntityNotFoundException("Department not found with id: " + departmentId);
         }
 
+        Custom custom = customRepository.getReferenceById(customId);
         Department department = departmentRepository.getReferenceById(departmentId);
-        if (department == null) {
-            throw new RuntimeException("Department not found with id: " + departmentId);
-        }
 
         custom.setDepartment(department);
         customRepository.save(custom);
